Move cleaning-function to coordinator.py
