+++ homeassistant/components/plugwise/util.py
@@ -6,12 +6,8 @@ from typing import Any, Concatenate, ParamSpec, TypeVar
 
 from plugwise.exceptions import PlugwiseException
 
-from homeassistant.config_entries import ConfigEntry
-from homeassistant.core import HomeAssistant, callback
 from homeassistant.exceptions import HomeAssistantError
-from homeassistant.helpers import entity_registry as er
 
-from .const import DOMAIN, LOGGER
 from .entity import PlugwiseEntity
 
 _PlugwiseEntityT = TypeVar("_PlugwiseEntityT", bound=PlugwiseEntity)
@@ -41,33 +37,3 @@ def plugwise_command(
             await self.coordinator.async_request_refresh()
 
     return handler
-
-
-@callback
-def _async_cleanup_registry_entries(
-    hass: HomeAssistant, entry: ConfigEntry, entry_id: str
-) -> None:
-    """Remove extra entities that are no longer part of the integration."""
-    entity_registry = er.async_get(hass)
-    current_unique_ids = hass.data[DOMAIN][entry_id].current_unique_ids
-
-    existing_entries = er.async_entries_for_config_entry(entity_registry, entry_id)
-    entities = {
-        (entity.domain, entity.unique_id): entity.entity_id
-        for entity in existing_entries
-    }
-
-    extra_entities = set(entities.keys()).difference(current_unique_ids)
-    if not extra_entities:
-        return
-
-    for entity in extra_entities:
-        LOGGER.debug("HOI entity: %s", entities[entity])
-        if entity_registry.async_is_registered(entities[entity]):
-            entity_registry.async_remove(entities[entity])
-
-    LOGGER.debug(
-        ("Clean-up of Plugwise entities: %s entities removed for config entry %s"),
-        len(extra_entities),
-        entry_id,
-    )
